Moved task utilities into their own subpackage and split them back into their own files
