+++ src/iarc7_motion/iarc_tasks/track_roomba_task.py
@@ -6,13 +6,10 @@ import tf2_ros
 import tf2_geometry_msgs
 import threading
 
-from geometry_msgs.msg import TwistStamped
-from geometry_msgs.msg import PointStamped
-from geometry_msgs.msg import Point
+from geometry_msgs.msg import TwistStamped, PointStamped, Point
 from iarc7_msgs.msg import OdometryArray
 from nav_msgs.msg import Odometry
-from geometry_msgs.msg import Vector3Stamped
-from geometry_msgs.msg import Vector3
+from geometry_msgs.msg import Vector3Stamped, Vector3
 
 from .abstract_task import AbstractTask
 from iarc_tasks.task_states import (TaskRunning,
@@ -23,7 +20,9 @@ from iarc_tasks.task_states import (TaskRunning,
 from iarc_tasks.task_commands import (VelocityCommand,
                                       NopCommand)
 
-from task_utilities import HeightHolder, HeightSettingsChecker, AccelerationLimiter
+from task_utilities.height_holder import HeightHolder
+from task_utilities.height_settings_checker import HeightSettingsChecker
+from task_utilities.acceleration_limiter import AccelerationLimiter
 
 class TrackObjectTaskState(object):
     init = 0
